requirements.txt
disnake
requests
python-dotenv
os